requirements.txt
pywebview
tk
meta-ai-api
pyinstaller